move warden and sculk directories into more logical places
