+++ src/main/java/net/frozenblock/wilderwild/mixin/block/sculk/SculkSpreaderChargeCursorMixin.java
@@ -16,7 +16,7 @@
  * along with this program; if not, see <https://www.gnu.org/licenses/>.
  */
 
-package net.frozenblock.wilderwild.mixin.sculk;
+package net.frozenblock.wilderwild.mixin.block.sculk;
 
 import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
 import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
