+++ src/main/java/net/frozenblock/wilderwild/mixin/entity/warden/SonicBoomMixin.java
@@ -16,7 +16,7 @@
  * along with this program; if not, see <https://www.gnu.org/licenses/>.
  */
 
-package net.frozenblock.wilderwild.mixin.warden;
+package net.frozenblock.wilderwild.mixin.entity.warden;
 
 import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
 import com.llamalad7.mixinextras.sugar.Share;
