+++ src/main/java/net/frozenblock/wilderwild/mixin/entity/warden/WardenMixin.java
@@ -16,7 +16,7 @@
  * along with this program; if not, see <https://www.gnu.org/licenses/>.
  */
 
-package net.frozenblock.wilderwild.mixin.warden;
+package net.frozenblock.wilderwild.mixin.entity.warden;
 
 import com.llamalad7.mixinextras.injector.ModifyReturnValue;
 import net.frozenblock.wilderwild.config.EntityConfig;
