+++ src/main/resources/wilderwild.mixins.json
@@ -64,23 +64,23 @@
     "projectile.ThrownEnderpearlMixin",
     "projectile.ThrownExperienceBottleMixin",
     "projectile.ThrownPotionMixin",
-    "sculk.CalibratedSculkSensorBlockEntityMixin",
-    "sculk.CalibratedSculkSensorBlockMixin",
-    "sculk.SculkBlockMixin",
-    "sculk.SculkSensorBlockEntityMixin",
+    "block.sculk.CalibratedSculkSensorBlockEntityMixin",
+    "block.sculk.CalibratedSculkSensorBlockMixin",
+    "block.sculk.SculkBlockMixin",
+    "block.sculk.SculkSensorBlockEntityMixin",
     "sculk.SculkSensorBlockEntityMixin$VibrationUserMixin",
-    "sculk.SculkSensorBlockMixin",
-    "sculk.SculkShriekerBlockEntityMixin",
+    "block.sculk.SculkSensorBlockMixin",
+    "block.sculk.SculkShriekerBlockEntityMixin",
     "sculk.SculkShriekerBlockEntityMixin$VibrationUserMixin",
-    "sculk.SculkShriekerBlockMixin",
-    "sculk.SculkSpreaderChargeCursorMixin",
-    "sculk.SculkVeinBlockMixin",
-    "sculk.SculkVeinSpreaderConfigMixin",
-    "warden.LivingEntityMixin",
-    "warden.SonicBoomMixin",
-    "warden.WardenMixin",
+    "block.sculk.SculkShriekerBlockMixin",
+    "block.sculk.SculkSpreaderChargeCursorMixin",
+    "block.sculk.SculkVeinBlockMixin",
+    "block.sculk.SculkVeinSpreaderConfigMixin",
+    "entity.warden.LivingEntityMixin",
+    "entity.warden.SonicBoomMixin",
+    "entity.warden.WardenMixin",
     "warden.WardenMixin$VibrationUserMixin",
-    "warden.WardenSwimMixin",
+    "entity.warden.WardenSwimMixin",
     "worldgen.biome.OverworldBiomeBuilderMixin",
     "worldgen.structure.SwampHutPieceMixin",
     "worldgen.tree.AbstractTreeGrowerMixin",
